Move auth.slice
