+++ src/features/board/navbar.tsx
@@ -3,7 +3,7 @@ import Link from "next/dist/client/link";
 import { AnchorLink } from "src/features/shared-styles";
 import { path } from "src/utils/url/drello-web";
 import { useDispatch } from "react-redux";
-import { signout } from "src/redux/auth.slice";
+import { signout } from "src/features/auth/auth.slice";
 
 const Main = styled.nav`
   display: grid;
