+++ src/redux/user.slice.ts
@@ -1,6 +1,6 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import { RootState } from "src/redux/root";
-import { signin, signout } from "src/redux/auth.slice";
+import { signin, signout } from "src/features/auth/auth.slice";
 import { getUser } from "src/features/user/user.api";
 import { User } from "src/types/user.g";
